app/requirements.txt
Flask
mysql-connector
simplejson
Authlib==0.14.1
requests
